# docs/openapi-v3.yml
openapi: 3.0.3
info:
  title: Yanki Wallet Service API
  version: v3
  description: Mobile wallet management and payments - Project III.

paths:
  /wallets:
    post:
      summary: Create Yanki wallet
      description: Creates a mobile wallet without requiring the user to be a bank customer.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/YankiWalletCreateRequest'
      responses:
        '201':
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YankiWalletResponse'

  /wallets/{id}:
    get:
      summary: Get wallet by ID
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Wallet retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YankiWalletResponse'

  /wallets/by-phone/{phoneNumber}:
    get:
      summary: Get wallet by phone number
      parameters:
        - in: path
          name: phoneNumber
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Wallet retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/YankiWalletResponse'

  /wallets/{id}/payments/send:
    post:
      summary: Send payment to another Yanki wallet
      description: Sends money using only the recipient's phone number.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/YankiSendPaymentRequest'
      responses:
        '201':
          description: Payment sent

  /wallets/{id}/payments/receive:
    post:
      summary: Receive payment from another Yanki wallet
      description: Registers an incoming payment.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/YankiReceivePaymentRequest'
      responses:
        '201':
          description: Payment received

  /wallets/{id}/link-debit-card:
    post:
      summary: Link Yanki wallet to a debit card
      description: Associates the wallet with a bank debit card so payments affect the main account.
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/YankiLinkDebitCardRequest'
      responses:
        '200':
          description: Debit card linked

components:
  schemas:
    YankiWalletResponse:
      type: object
      properties:
        id: { type: string }
        documentType: { $ref: '#/components/schemas/DocumentType' }
        documentNumber: { type: string }
        phoneNumber: { type: string }
        imei: { type: string }
        email: { type: string }
        linkedDebitCardId: { type: string }
        status:
          type: string
          enum: [ACTIVE, INACTIVE]

    YankiWalletCreateRequest:
      type: object
      required: [documentType, documentNumber, phoneNumber, imei, email]
      properties:
        documentType: { $ref: '#/components/schemas/DocumentType' }
        documentNumber: { type: string }
        phoneNumber: { type: string }
        imei: { type: string }
        email: { type: string }

    YankiSendPaymentRequest:
      type: object
      required: [toPhoneNumber, amount]
      properties:
        toPhoneNumber: { type: string }
        amount: { type: number }
        description: { type: string }

    YankiReceivePaymentRequest:
      type: object
      required: [fromPhoneNumber, amount]
      properties:
        fromPhoneNumber: { type: string }
        amount: { type: number }
        description: { type: string }

    YankiLinkDebitCardRequest:
      type: object
      required: [debitCardId]
      properties:
        debitCardId: { type: string }

    DocumentType:
      type: string
      enum: [DNI, RUC, PASSPORT]